RF:  move _fiducial_coords() -> viz._3d
