+++ mne/gui/_file_traits.py
@@ -20,9 +20,10 @@ from ..bem import read_bem_surfaces
 from ..io.constants import FIFF
 from ..io import read_info, read_fiducials
 from ..surface import read_surface
-from ..coreg import (_is_mri_subject, _mri_subject_has_bem, _fiducial_coords,
+from ..coreg import (_is_mri_subject, _mri_subject_has_bem,
                      create_default_subject)
 from ..utils import get_config, set_config
+from ..viz._3d import _fiducial_coords
 
 
 fid_wildcard = "*.fif"
